+++ Sources/BraintreeDataCollector/BTDataCollector.swift
@@ -91,10 +91,10 @@ import BraintreeCore
     /// Use the return value on your server, e.g. with `Transaction.sale`. or in
     /// client side requests such as PayPal, Venmo, or Local Payment Methods.
     /// - Parameters:
-    ///   - riskCorrelationID: A risk correlation ID to associate with this device data collection
+    ///   - riskCorrelationID: Optional: A risk correlation ID to associate with this device data collection
     ///   - completion: A completion block that returns either a device data string or an error with the failure reason. Retries are recommended on failure.
     @objc public func collectDeviceDataOnSuccess(
-        riskCorrelationID: String,
+        riskCorrelationID: String? = nil,
         _ completion: @escaping (String?, Error?) -> Void
     ) {
         fetchConfiguration { configuration, error in
@@ -114,19 +114,17 @@ import BraintreeCore
                 magnesSource: .BRAINTREE
             )
 
-            _ = try? MagnesSDK.shared().collectAndSubmit(
-                withPayPalClientMetadataId: riskCorrelationID,
+            var magnesResult: MagnesResult?
+            magnesResult = try? MagnesSDK.shared().collectAndSubmit(
+                withPayPalClientMetadataId: riskCorrelationID ?? "",
                 withAdditionalData: [:]
-            ) { status, clientMetadataID in
+            ) { status, _ in
                 switch status {
                 case .success:
-                    guard let clientMetadataID else {
-                        completion(nil, BTDataCollectorError.callbackSubmitError)
-                        return
-                    }
+                    let clientMetadataID = magnesResult?.getPayPalClientMetaDataId() ?? ""
 
                     let data: [String: String] = ["correlation_id": clientMetadataID]
-                    
+
                     guard let jsonData = try? JSONSerialization.data(withJSONObject: data) else {
                         completion(nil, BTDataCollectorError.jsonSerializationFailure)
                         return
@@ -162,10 +160,10 @@ import BraintreeCore
     /// client side requests such as PayPal, Venmo, or Local Payment Methods.
     /// - Parameters:
     ///  - completion: A completion block that returns either a device data string or an error with the failure reason. Retries are recommended on failure.
-    ///  - riskCorrelationID: A risk correlation ID to associate with this device data collection
+    ///  - riskCorrelationID: Optional: A risk correlation ID to associate with this device data collection
     /// - Returns: A device data string that should be passed into server-side calls, such as `Transaction.sale`.
     /// - Throws: An `Error` describing the failure or timeout. Merchants should retry on error to ensure coverage.
-    public func collectDeviceDataOnSuccess(riskCorrelationID: String) async throws -> String {
+    public func collectDeviceDataOnSuccess(riskCorrelationID: String? = nil) async throws -> String {
         try await withCheckedThrowingContinuation { continuation in
             collectDeviceDataOnSuccess(riskCorrelationID: riskCorrelationID) { deviceData, error in
                 if let error {